backend/api_foodgram/requirements.txt
requests==2.26.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-import-export==2.7.1
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
python-decouple==3.6
Pillow==9.2.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-extra-fields==3.0.2
isort==5.10.1
django-cors-headers==3.13.0
reportlab == 3.6.12
django-crispy-forms==1.14.0
python-dotenv==0.21.0
